c++ schenanigans

+++ src/axon_link/src/testSerial.cpp
@@ -53,7 +53,7 @@ void commandSend(unsigned char a, unsigned char b, unsigned char c){
 
 		for (j = 0; j < sizeof(reply); j ++) {
 
-			cout << reply[j] << endl;
+			cout << (int)reply[j] << endl;
 			cout <<"this is reply " << j << endl;
 			}
 		}
